fix circular deps
